new RCP webhook (#38)
We are introducing a new webhook variant alongside the current version. 
The new webhook uses the `rcp.phase.update` action instead of `rcp.status.update`. 
We strongly advise using the new version, as the old webhook may not provide sufficient detail about the receipt status. For example, an order may be reported as closed while customs activities are still pending. In such cases, stock might not yet be available for ordering, which could result in rejected orders due to unavailable inventory.